added caching for reversegeocodings. #949 renaming some functions and arguments to make em kidfriendly.
